# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: parking API
  description: |
    API REST de gestión de plazas de parking de ALEATICA.
    Autenticación por cookie de sesión (`parking_SESSION`). Fase 1: login local.
    Fase 2: SSO ALEATICA (la landing autentica, parking autoriza).
  version: 1.0.0
  contact:
    name: Equipo parking
    email: "_[pendiente]_"
servers:
  - url: http://localhost:8080/parking-api/api/v1
    description: Desarrollo local
  - url: https://parking.aleatica.com/parking-api/api/v1
    description: Producción
...
security:
  - sessionCookie: []
paths:
...
  /desks:
    get:
      tags: [Desks]
      operationId: listDesks
      summary: Lista puestos (ADMIN o EMPLOYEE)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: active
          in: query
          required: false
          schema: { type: boolean }
      responses:
        '200':
          description: Página de puestos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDesk'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
    post:
      tags: [Desks]
      operationId: createDesk
      summary: Crea un puesto (ADMIN)
      description: '`number` único en el rango 1-65 (400 fuera de rango, 409 en colisión).'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeskCreate'
      responses:
        '201':
          description: Puesto creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Desk'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
  /desks/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Desks]
      operationId: getDesk
      summary: Detalle de un puesto (ADMIN o EMPLOYEE)
      responses:
        '200':
          description: Puesto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Desk'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
    put:
      tags: [Desks]
      operationId: updateDesk
      summary: Modifica categoría y coordenadas de un puesto (ADMIN)
      description: 'El `number` es inmutable. Coordenadas en el rango 0-100 (400 fuera de rango).'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeskUpdate'
      responses:
        '200':
          description: Puesto actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Desk'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /desks/{id}/activation:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    patch:
      tags: [Desks]
      operationId: setDeskActivation
      summary: Activa o desactiva un puesto (ADMIN)
      description: Un puesto inactivo deja de aparecer como disponible.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeskActivationRequest'
      responses:
        '200':
          description: Estado del puesto actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Desk'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }

  # ---------------------------------------------------------------------------
  # Floor Plan
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: parking_SESSION
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
    PageParam:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    SizeParam:
      name: size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
    BadRequest:
      description: Datos inválidos
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Sin permisos para la operación
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Conflicto de estado o unicidad
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ServerError:
      description: Error interno
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
...
    ApiError:
      type: object
      required: [error, message, timestamp]
      properties:
        error:
          type: string
          description: Código corto de error.
        message:
          type: string
        fields:
          type: object
          additionalProperties: { type: string }
          description: Errores por campo (validación).
        timestamp:
          type: string
          format: date-time

    # ---- Pagination ----
    PageMeta:
      type: object
      properties:
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        size: { type: integer }
        number: { type: integer }
        first: { type: boolean }
        last: { type: boolean }

    # ---- Auth ----
...
    DeskCategory:
      type: string
      enum: [STANDARD, EXECUTIVE]
      description: >
        Categoría del puesto. Determina la auto-asignación por rango del empleado (change
        `desk-auto-assignment`, `POST /requests` DESK + AUTOMATIC sin `resourceId`): EXECUTIVE
        es exclusivo de las categorías altas (CEO, CONSEJO, DIRECTOR_N1, DIRECTOR_N2), que
        caen a STANDARD si no hay ninguno libre; el resto de categorías solo candidata
        STANDARD. Fuera de la auto-asignación (asignación fija, elección explícita), ambas
        categorías se comportan igual.
    Desk:
      type: object
      properties:
        id: { type: integer, format: int64 }
        number: { type: integer, minimum: 1, maximum: 65 }
        label: { type: string, description: 'Etiqueta derivada del número (p. ej. D-12)' }
        category: { $ref: '#/components/schemas/DeskCategory' }
        coordX: { type: number, format: double, minimum: 0, maximum: 100 }
        coordY: { type: number, format: double, minimum: 0, maximum: 100 }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
    DeskCreate:
      type: object
      required: [number, category]
      properties:
        number: { type: integer, minimum: 1, maximum: 65 }
        category: { $ref: '#/components/schemas/DeskCategory' }
        coordX:
          type: number
          format: double
          minimum: 0
          maximum: 100
          description: Opcional; si se omite se centra (50).
        coordY:
          type: number
          format: double
          minimum: 0
          maximum: 100
          description: Opcional; si se omite se centra (50).
    DeskUpdate:
      type: object
      required: [category]
      description: 'El número es inmutable. Coordenadas opcionales: si se omiten, se conservan.'
      properties:
        category: { $ref: '#/components/schemas/DeskCategory' }
        coordX: { type: number, format: double, minimum: 0, maximum: 100 }
        coordY: { type: number, format: double, minimum: 0, maximum: 100 }
    DeskActivationRequest:
      type: object
      required: [active]
      properties:
        active: { type: boolean }
    PageDesk:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/Desk' }
...